all: move pkgs used only by agent inside agent
